# swagger.yaml
swagger: "2.0"
info:
  description: "Learning NestJS for building scalable Node.js server solutions."
  version: "1.0.0"
  title: "NestJS Fundamentals"
host: "localhost"
basePath: "/v1"
...
schemes:
- "https"
- "http"
paths:
...
  /distillery:
    post:
      tags:
      - "distillery"
      summary: "Create a distillery"
      operationId: "createDistillery"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created distillery object"
        required: true
        schema:
          $ref: "#/definitions/Distillery"
      responses:
        default:
          description: "successful operation"
  /distillery/{id}:
    get:
      tags:
      - "distillery"
      summary: "Get distillery by id"
      description: ""
      operationId: "getDistilleryById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "number"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Distillery"
        400:
          description: "Invalid distillery id"
        404:
          description: "distillery not found"
    put:
      tags:
      - "distillery"
      summary: "Updated distillery"
      operationId: "updateDistilleryById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "number"
      - in: "body"
        name: "body"
        description: "Updated distillery object"
        required: true
        schema:
          $ref: "#/definitions/Distillery"
      responses:
        400:
          description: "Invalid distillery id"
        404:
          description: "distillery not found"
    delete:
      tags:
      - "distillery"
      summary: "Delete distillery"
      operationId: "deleteDistilleryById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "number"
      responses:
        400:
          description: "Invalid distillery id"
        404:
          description: "distillery not found"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "https://localhost/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify scotches in your account"
      read:pets: "read your scotches"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Distillery:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      region:
        $ref: "#/definitions/Region"
      country:
        type: "string"
      city:
        type: "string"
  Region:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
...